Added file w/ DWT-related functions.
evalComplexity() and visualize() are the only two functions that should be called outside of this file.

evalComplexity: im and mask are fed from load_helper() in utils. "thrPecentile", "levels" and "wavelet" can be set by the user so there should be an option somewhere in the parser.

visualize: no parameters needed. Just add an option when launching the DWT method to visualize the coefficients (levels and mWavelet would be the same ones used in evalComplexity).

Everything else in this file is filler to make those two functions run. Don't ask me why I use two different image resize functions. It just works TM.